requirements.txt
PyYAML
RPi.GPIO
gpiozero
lirc
netifaces
wavio
soundfile
